stake-pool: Split up more tests to help CI, avoid warping
